backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv>=1.0
google-genai
Pillow
sqlalchemy>=2.0
psycopg[binary]>=3.1
feedparser>=6.0
requests>=2.31